+++ source/contracts/trading/ShareToken.sol
@@ -12,7 +12,7 @@ import 'reporting/IMarket.sol';
 contract ShareToken is DelegationTarget, ITyped, Initializable, VariableSupplyToken, IShareToken {
 
     string constant public name = "Shares";
-    uint256 constant public decimals = 0;
+    uint8 constant public decimals = 0;
     string constant public symbol = "SHARE";
 
     IMarket private market;
